api_student_management/api/requirements.txt
django
django-rest-framework
pillow
psycopg2
django-cors-headers
django_filter